tests/wpt/tests/html/dom/render-blocking/element-render-blocking-019.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Media attribute changes in the body to not apply</title>

<link id=link rel=expect href="#last" blocking="render" media="(min-width: 10px)">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "changing media to non-matching in the body makes it non blocking");
</script>
</head>
<body>
<script>
link.media = "(max-width: 10px)";
</script>

  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
